devices: pci: provide bar_index and offset in read/write_bar()
The PciDevice::read_bar() and PciDevice::write_bar() functions are now
passed a BAR index and offset within that BAR, rather than a guest
physical address. This simplifies the job of each device implementation
since it is no longer necessary to look up which BAR corresponds to the
provided address (this is handled in generic PciDevice code now).

Providing the address information in this way also ensures devices will
work even in the face of BAR reprogramming; there is no way to compare
against a stale BAR address, since that information is not provided to
the read_bar()/write_bar() callbacks anymore.

BUG=None
TEST=tools/dev_container tools/presubmit
TEST=Boot Crostini and ARCVM on brya

Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/4564155
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>

+++ devices/src/platform/vfio_platform.rs
@@ -38,7 +38,7 @@ use crate::IrqLevelEvent;
 use crate::Suspendable;
 
 struct MmioInfo {
-    index: u32,
+    index: usize,
     start: u64,
     length: u64,
 }
@@ -181,7 +181,7 @@ impl VfioPlatformDevice {
         Ok(ranges)
     }
 
-    fn region_mmap(&self, index: u32, start_addr: u64) -> Vec<MemoryMapping> {
+    fn region_mmap(&self, index: usize, start_addr: u64) -> Vec<MemoryMapping> {
         let mut mem_map: Vec<MemoryMapping> = Vec::new();
         if self.device.get_region_flags(index) & VFIO_REGION_INFO_FLAG_MMAP != 0 {
             let mmaps = self.device.get_region_mmap(index);
